requirements_Elysia.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.11.0
attrs==25.3.0
autopep8==2.3.2
bcrypt==4.3.0
beautifulsoup4==4.13.3
Brotli==1.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
chinese_calendar==1.10.0
click==8.1.8
colorama==0.4.6
crypto==1.4.1
cryptography==44.0.2
dashscope==1.22.2
distro==1.9.0
exceptiongroup==1.2.2
fake-useragent==2.1.0
fastapi==0.115.11
ffmpeg-python==0.2.0
filelock==3.18.0
frozenlist==1.5.0
fsspec==2025.3.0
future==1.0.0
gradio_client==1.8.0
graiax-silkcoder==0.3.6
greenlet==3.1.1
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.29.3
hyperframe==6.1.0
idna==3.10
imageio-ffmpeg==0.6.0
iso8601==2.1.0
Jinja2==3.1.6
jiter==0.9.0
lazy-object-proxy==1.10.0
loguru==0.7.3
lunardate==0.2.2
Markdown==3.7
MarkupSafe==3.0.2
multidict==6.2.0
Naked==0.1.32
nonebot-adapter-qq==1.6.2
nonebot-plugin-htmlrender==0.6.2
nonebot2==2.4.2
numpy==2.2.4
openai==1.68.2
opuslib_next==1.1.2
outcome==1.3.0.post0
packaging==24.2
paramiko==3.5.1
pillow==11.1.0
playwright==1.51.0
propcache==0.3.0
psutil==7.0.0
pycodestyle==2.12.1
pycparser==2.22
pycryptodome==3.22.0
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
pyee==12.1.1
PyExecJS==1.5.1
Pygments==2.19.1
pygtrie==2.5.0
pymdown-extensions==10.14.3
PyNaCl==1.5.0
pypika-tortoise==0.5.0
PySocks==1.7.1
python-dotenv==1.0.1
python-markdown-math==0.8
pytz==2025.1
PyYAML==6.0.2
qrcode==8.0
requests==2.32.3
ruamel.yaml.clib==0.2.12
selenium==4.30.0
shellescape==3.8.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.39
starlette==0.46.1
tenacity==9.0.0
tortoise-orm==0.24.2
tqdm==4.67.1
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3.1
urllib3==2.3.0
uvicorn==0.34.0
watchfiles==1.0.4
websocket-client==1.8.0
websockets==15.0.1
win32_setctime==1.2.0
wsproto==1.2.0
yarl==1.18.3
zhdate==0.1
